solve all warnings
