+++ docs-src/src/components/Content/TagArea.js
@@ -16,13 +16,14 @@ class TagArea extends Component{
     render(){
         console.log(this.props.tags);
         let tags = this.props.tags;
-        if(tags == undefined){
+        if(tags === undefined || tags === null){
             return (<div></div>);
         }else{
-            return (<div className="tags">🔖
+            return (<div className="tags">
+                <span role="img" aria-label="tags">🔖</span>
                 {
                     tags.map((v, i) => {
-                        if(i != 0){
+                        if(i !== 0){
                             return (<a href={"/docs/" + v}>{v}</a>)
                         }else{
                             return (<a href={"/docs/" + v}>{v}</a>)
