--- docs-src/src/components/Content/index.js
@@ -1,4 +1,0 @@
-import List from './List'
-import Right from './Right'
-
-export {List,Right}
