+++ docs-src/src/components/Layouts/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import {Container,Typography,Grid} from '@material-ui/core';
+import { Typography,Grid} from '@material-ui/core';
 import logo from '../source/logo.png'
