+++ docs-src/src/components/Layouts/index.js
@@ -1,5 +1,4 @@
 import Header from './Header'
 import Footer from './Footer'
-import Content from './Content'
 
-export {Header,Footer,Content}
+export {Header,Footer}
