--- docs-src/src/components/catagory_list.js
@@ -1,21 +1,0 @@
-import React from "react"
-import {Container, ListGroup} from "react-bootstrap"
-
-export default function Catagory(props){
-    console.log(props.tags_list);
-    const tags_list = props.tags_list;
-    return (
-<Container>
-          <h3>Category</h3>
-          <ListGroup>
-          {
-            tags_list.map((v, i) => {
-                if(v != null)
-                return (<ListGroup.Item key={i}><a href={'/docs/'+v}>{v}</a></ListGroup.Item>);
-                else
-                  return "";
-            })
-          }
-          </ListGroup>
-        </Container>);
-}
