+++ docs-src/src/components/pagination.js
@@ -3,6 +3,7 @@ import { Grid } from '@material-ui/core';
 import React from "react"
 export default function Page(props){
     // console.log(props)
+    // 分页的下边栏
     return (
     <Grid  container direction="row" justify="center" alignItems="center">
         <Pagination count={props.totalPage} defaultPage={props.defaultPage} onChange={props.onChange}/>
